message-service/requirements.txt
# FastAPI and web framework dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database dependencies
motor==3.3.2
beanie==1.23.6
pymongo==4.6.0

# Redis dependencies
redis==5.0.1
redis[hiredis]==5.0.1

# Authentication and security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Environment and configuration
python-decouple==3.8
pydantic-settings==2.1.0

# HTTP client for external services
httpx==0.25.2

# Logging and monitoring
structlog==23.2.0
prometheus-client==0.19.0

# Content processing
bleach==6.1.0

# Development dependencies (will be in requirements-dev.txt)
# pytest==7.4.3
# pytest-asyncio==0.21.1
# pytest-cov==4.1.0
# black==23.11.0
# isort==5.12.0
# flake8==6.1.0
# mypy==1.7.1